requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For voice output
pyttsx3>=2.90

# Optional: For voice input
SpeechRecognition>=3.10.0
pyaudio>=0.2.13  # May require system-level audio libraries